+++ ThrEd4/globals.cpp
@@ -29,13 +29,13 @@
 
 namespace fs = std::experimental::filesystem;
 
-unsigned   ActiveColor = 0; // active color selector
-unsigned   ActiveLayer = 0; // active layer
+uint32_t   ActiveColor = 0; // active color selector
+uint32_t   ActiveLayer = 0; // active layer
 fRECTANGLE AllItemsRect;    // rectangle enclosing all forms and stitches
 
 std::vector<fPOINT>* AngledFormVertices; // form formOrigin data for angle fills
 
-unsigned  AppliqueColor = 15; // underlay color
+uint32_t  AppliqueColor = 15; // underlay color
 LPWSTR*   ArgList;            // command line argument array
 HCURSOR   ArrowCursor;        // arrow
 fs::path* AuxName;            //
@@ -43,25 +43,25 @@ fs::path* AuxName;            //
 std::vector<BSEQPNT>* BSequence; //
 
 HBRUSH   BackgroundBrush;      // background color brush
-unsigned BeanCount;            // number of stitches added by convert to bean
+uint32_t BeanCount;            // number of stitches added by convert to bean
 HDC      BitmapDC;             // bitmap device context
 RECT     BitmapDstRect;        // stitch window destination rectangle for zooomed view
-unsigned BitmapHeight;         // bitmap height
+uint32_t BitmapHeight;         // bitmap height
 POINT    BitmapPoint;          // a point on the bitmap
 dPOINT   BitmapSizeinStitches; // bitmap end points in stitch points
 RECT     BitmapSrcRect;        // bitmap source rectangle for zoomed view
 dPOINT   BmpStitchRatio;       // bitmap to stitch hoop ratios
-unsigned BitmapWidth;          // bitmap width
+uint32_t BitmapWidth;          // bitmap width
 float    BorderWidth = BRDWID; // border width for satin borders
-unsigned ButtonHeight;         // button height
-unsigned ButtonWidth;          // button width
-unsigned ButtonWidthX3;        // button width times 3
+uint32_t ButtonHeight;         // button height
+uint32_t ButtonWidth;          // button width
+uint32_t ButtonWidthX3;        // button width times 3
 
 std::vector<HWND>* ButtonWin; // button windows
 
 float    ButtonholeCornerLength = IBFCLEN; // buttonhole corner length
 HWND     CancelButton;                     // cancel button
-unsigned Clip = 0;                         // pcs format clipboard handle
+uint32_t Clip = 0;                         // pcs format clipboard handle
 
 std::vector<fPOINTATTR>* ClipBuffer; //
 
@@ -75,16 +75,16 @@ fRECTANGLE ClipRect;       // clipboard rectangle
 FLSIZ      ClipRectSize;   // clipboard rectangle size
 CLPSTCH*   ClipStitchData; // for pcs clipboard data
 
-unsigned ClipTypeMap = MCLPF | MVCLPF | MHCLPF | MANGCLPF; // for checking if a fill is a clipboard fill
+uint32_t ClipTypeMap = MCLPF | MVCLPF | MHCLPF | MANGCLPF; // for checking if a fill is a clipboard fill
 
-unsigned int ClosestFormToCursor;    // closest form to the cursor
-unsigned     ClosestPointIndex;      // index of closest point
-unsigned int ClosestVertexToCursor;  // formOrigin closest to the cursor
-int*         ColorBarSize;           // Color bar width scaled for DPI
+uint32_t ClosestFormToCursor;    // closest form to the cursor
+uint32_t     ClosestPointIndex;      // index of closest point
+uint32_t ClosestVertexToCursor;  // formOrigin closest to the cursor
+int32_t*         ColorBarSize;           // Color bar width scaled for DPI
 HCURSOR      CrossCursor;            // cross
-unsigned int CurrentFormGuides;      // guide index in the currently selected form
-unsigned int CurrentFormGuidesCount; // number of guides in the currently selected form
-unsigned int CurrentVertexIndex;     // points in the currently selected form
+uint32_t CurrentFormGuides;      // guide index in the currently selected form
+uint32_t CurrentFormGuidesCount; // number of guides in the currently selected form
+uint32_t CurrentVertexIndex;     // points in the currently selected form
 
 std::vector<HWND>* DefaultColorWin; // default color windows
 
@@ -101,7 +101,7 @@ std::vector<POINT>* FormControlPoints; // form control rectangle in pixel coordi
 HWND                 FormDataSheet = nullptr; // form data sheet
 std::vector<POINT>*  FormLines;               // used in the form drawing routines
 std::vector<FRMHED>* FormList;                // a list of form headers
-unsigned int         FormMenuChoice = 0u;     // data type for form data form numerical entry
+uint32_t         FormMenuChoice = 0u;     // data type for form data form numerical entry
 fPOINT               FormMoveDelta;           // offset for moving forms
 
 std::wstring* FormOnOff;
@@ -114,9 +114,9 @@ std::vector<fPOINT>* FormVertices; // form points
 
 HWND      GeneralNumberInputBox; // general number input box
 HPEN      GridPen;               // pen for stitch grid
-unsigned  GroupEndStitch;        // higher end of selected stitches
-unsigned  GroupStartStitch;      // lower end of selected stitches
-unsigned  GroupStitchIndex;      // last point selected in group
+uint32_t  GroupEndStitch;        // higher end of selected stitches
+uint32_t  GroupStartStitch;      // lower end of selected stitches
+uint32_t  GroupStitchIndex;      // last point selected in group
 fs::path* HomeDirectory;         // directory from which thred was executed
 double    HorizontalLength;      // horizontal length of a clipboard fill
 float     HorizontalLength2;     // horizontal length of a clipboard fill/2
@@ -144,18 +144,18 @@ MENUITEMINFO* MenuInfo;
 float        MinStitchLength = MINSIZ * PFAFGRAN; // minimum stitch size
 MSG          Msg;                                 // main message loop message
 wchar_t      MsgBuffer[MSGSIZ];                   // for user messages
-unsigned     MsgIndex;                            // pointer to the message buffer
+uint32_t     MsgIndex;                            // pointer to the message buffer
 RECT         MsgRect;                             // rectangle containing the text message
 HWND         MsgWindow = nullptr;                 // message window
 HPEN         MultiFormPen;                        // multiple selected forms pen
-unsigned     NearestCount;                        // number of boxes selected
-unsigned int NewFormVertexCount;                  // points in the new form
+uint32_t     NearestCount;                        // number of boxes selected
+uint32_t NewFormVertexCount;                  // points in the new form
 HWND         OKButton;                            // ok button
 
 std::vector<fPOINT>* OSequence; // temporary storage for sequencing
 
 bool         OutLineEverySelectedForm = false; // When selecting multiple forms, should we outline every form?
-unsigned int OutputIndex;                      // output pointer for sequencing
+uint32_t OutputIndex;                      // output pointer for sequencing
 
 std::vector<fPOINT>* OutsidePointList; // list of outside outline points for satin or clipboard fills
 std::vector<fPOINT>* OutsidePoints;    // pointer to the list of outside outline points
@@ -164,30 +164,30 @@ char           PCSBMPFileName[16];                // bitmap file name from pcs f
 PCSHEADER      PCSHeader;                         // pcs file header
 const wchar_t* PcdClipFormat   = L"PMust_Format"; //
 float          PicotSpacing    = IPICSPAC;        // space between border picots
-unsigned       PreferenceIndex = 0;               // index to the active preference window
-long           PreferenceWindowWidth;             // width of the preference window
+uint32_t       PreferenceIndex = 0;               // index to the active preference window
+int32_t           PreferenceWindowWidth;             // width of the preference window
 HWND           PreferencesWindow;                 // preferences window
-unsigned int   PreviousFormIndex;                 // previously selected form
-unsigned       PseudoRandomValue;                 // pseudo-random sequence register
+uint32_t   PreviousFormIndex;                 // previously selected form
+uint32_t       PseudoRandomValue;                 // pseudo-random sequence register
 fRECTANGLE     RotationRect;                      // rotation rectangle
 
 std::vector<POINT>* RubberBandLine; // points to form points to be moved
 
-unsigned int SatinEndGuide; // satin end guide for the currently selected form
+uint32_t SatinEndGuide; // satin end guide for the currently selected form
 
 std::vector<SATCON>* SatinGuides; // satin form connects
 
 RECT      scRct;      //
-unsigned* screenDPI;  // screen DPI
-int*      ScrollSize; // Scroll bar width scaled for DPI
+uint32_t* screenDPI;  // screen DPI
+int32_t*      ScrollSize; // Scroll bar width scaled for DPI
 
 std::vector<POINT>* SearchLine; // stitch select line
 
 HPEN     SelectAllPen;              // pen for drawing large boxes
 FRMHED*  SelectedForm;              // pointer to selected form
-unsigned SelectedFormControlVertex; // user selected form control formOrigin
+uint32_t SelectedFormControlVertex; // user selected form control formOrigin
 
-std::vector<unsigned int>* SelectedFormList;     // a list of selected forms
+std::vector<uint32_t>* SelectedFormList;     // a list of selected forms
 FORMVERTICES               SelectedFormVertices; // selected form vertices
 std::vector<POINT>*        SelectedFormsLine;    // line derived from the big rectangle
 
@@ -224,17 +224,17 @@ POINT      StitchWindowSize;               // size of the stitch window in pixel
 std::vector<std::wstring>* StringTable; //
 std::vector<fPOINT>*       TempPolygon; // temporary storage when user is entering a polygon;
 
-unsigned int        TextureIndex;                   // next textured fill point index
+uint32_t        TextureIndex;                   // next textured fill point index
 std::wstring*       TextureInputBuffer;             // texture fill number buffer
 std::vector<TXPNT>* TexturePointsBuffer;            // buffer for textured fill points
 HWND                ThreadSizeWin[16];              // thread size windows
-unsigned            ThrEdClip       = 0;            // thred format
+uint32_t            ThrEdClip       = 0;            // thred format
 const wchar_t*      ThrEdClipFormat = L"threditor"; //
 HINSTANCE           ThrEdInstance;                  // main instance handle
 HWND                ThrEdWindow;                    // main window handle
 fs::path*           ThrName;                        //
 POINT               ThredWindowOrigin;              // offset origin of the main window
-unsigned*           TraceBitmapData;                // trace bitmap data
+uint32_t*           TraceBitmapData;                // trace bitmap data
 HDC                 TraceDC;                        // trace device context
 
 boost::dynamic_bitset<>* TracedEdges; // detected edges of trace areas
@@ -247,7 +247,7 @@ EnumMap<UserFlag>  UserFlagMap(0);                       // for storage of persi
 HPEN               UserPen[16];                          // user color pens
 float              UserStitchLength = USESIZ * PFAFGRAN; // user selected stitch size
 std::vector<HWND>* ValueWindow;                          // data handles for the form data sheet
-unsigned int       VertexCount;                          // sides of the selected form to fill
+uint32_t       VertexCount;                          // sides of the selected form to fill
 float              VerticalRatio;                        // vertical ratio between the zoom window and the entire stitch space
 HWND               VerticalScrollBar;                    // vertical scroll bar
 fs::path*          WorkingFileName;                      //
